Create a Projection Matrix: RotMatPPO

Create a projection matrix using projection pursuit optimization (PPO).

Usage

RotMatPPO(
  X,
  y,
  model = "PPR",
  type = "i-classification",
  weights = NULL,
  dimProj = min(ceiling(length(y)^0.4), ceiling(ncol(X) * 2/3)),
  numProj = ifelse(dimProj == "Rand", sample(floor(ncol(X)/3), 1),
    ceiling(ncol(X)/dimProj)),
  catLabel = NULL,
  ...
)

Arguments

X

An n by d numeric matrix (preferable) or data frame.

y

A response vector of length n.

model

Model for projection pursuit (for details see PPO).

type

The criterion used for splitting the variable. 'i-classification': information gain (classification, default), 'g-classification': gini impurity index (classification) or 'regression': mean square error (regression).

weights

A vector of length same as data that are positive weights. (default NULL)

dimProj

Number of variables to be projected, dimProj=min(ceiling(n^0.4),ceiling(ncol(X)*2/3)) (default) or dimProj="Rand": random from 1 to ncol(X).

numProj

The number of projection directions, when dimProj="Rand" default numProj = sample(ceiling(ncol(X)/3),1) otherwise default numProj=ceiling(p0/dimProj).

catLabel

A category labels of class list in predictors. (default NULL, for details see Examples of ODT)

...

Used to handle superfluous arguments passed in using paramList.

Value

A random matrix to use in running ODT.

Variable: Variables to be projected.

Number: Number of projections.

Coefficient: Coefficients of the projection matrix.

See also

Examples

set.seed(220828)
X <- matrix(rnorm(1000), 100, 10)
y <- (rnorm(100) > 0) + 0
(RotMat <- RotMatPPO(X, y))
#>       Variable Number Coefficient
#>  [1,]        7      1  1.00000000
#>  [2,]        3      2  1.00000000
#>  [3,]        2      3  1.00000000
#>  [4,]       10      4  1.00000000
#>  [5,]        8      5 -0.63747080
#>  [6,]        7      5 -0.33498954
#>  [7,]        6      5  0.40102900
#>  [8,]        4      5  0.44271368
#>  [9,]       10      5 -0.35297779
#> [10,]        2      6 -0.22464801
#> [11,]        1      6  0.70400011
#> [12,]        5      6  0.16604464
#> [13,]        3      6 -0.04357413
#> [14,]        9      6  0.65149643
#> [15,]        1      7  0.62001403
#> [16,]        9      7  0.26056728
#> [17,]        8      7 -0.12300744
#> [18,]        4      7  0.40539513
#> [19,]        6      7  0.60680413
(RotMat <- RotMatPPO(X, y, dimProj = "Rand"))
#>       Variable Number Coefficient
#>  [1,]        9      1   1.0000000
#>  [2,]       10      2   1.0000000
#>  [3,]        3      3   1.0000000
#>  [4,]        1      4   1.0000000
#>  [5,]        2      5   1.0000000
#>  [6,]        4      6   0.3429718
#>  [7,]        1      6   0.4460246
#>  [8,]       10      6  -0.5441787
#>  [9,]        7      6  -0.3503986
#> [10,]        6      6   0.5143178
#> [11,]        3      7  -0.2794397
#> [12,]        8      7  -0.9591879
#> [13,]        2      7  -0.0432669
#> [14,]        8      8  -0.3423948
#> [15,]       10      8  -0.6147243
#> [16,]        6      8   0.5486710
#> [17,]        1      8   0.4295487
#> [18,]        7      8  -0.1390248
(RotMat <- RotMatPPO(X, y, dimProj = 6, numProj = 4))
#>       Variable Number Coefficient
#>  [1,]        7      1   1.0000000
#>  [2,]        6      2   1.0000000
#>  [3,]        5      3   1.0000000
#>  [4,]       10      4   1.0000000
#>  [5,]       10      5  -0.6616202
#>  [6,]        2      5   0.2554465
#>  [7,]        9      5   0.7049864
#>  [8,]        7      6  -0.7046850
#>  [9,]        1      6   0.7095203
#> [10,]        3      7  -0.6309069
#> [11,]        5      7   0.7758586
#> [12,]        6      8   0.5181236
#> [13,]        8      8  -0.7727649
#> [14,]        4      8   0.3665820
#> [15,]        5      9   0.2120148
#> [16,]        8      9  -0.4452757
#> [17,]        1      9   0.7179317
#> [18,]        9      9   0.4912772

# classification
data(seeds)
(PP <- RotMatPPO(seeds[, 1:7], seeds[, 8], model = "Log", type = "i-classification"))
#>       Variable Number Coefficient
#>  [1,]        2      1  1.00000000
#>  [2,]        3      2  1.00000000
#>  [3,]        1      3  1.00000000
#>  [4,]        2      4 -0.44768074
#>  [5,]        1      4 -0.11812580
#>  [6,]        6      4  0.10440721
#>  [7,]        7      4  0.88018599
#>  [8,]        4      5 -0.11263550
#>  [9,]        3      5 -0.99356630
#> [10,]        5      5  0.01180059
#> [11,]        3      6 -0.97305741
#> [12,]        7      6  0.08196076
#> [13,]        2      6 -0.20369703
#> [14,]        1      6  0.07035082
(PP <- RotMatPPO(seeds[, 1:7], seeds[, 8], model = "PPR", type = "i-classification"))
#>       Variable Number Coefficient
#>  [1,]        5      1  1.00000000
#>  [2,]        1      2  1.00000000
#>  [3,]        6      3  1.00000000
#>  [4,]        5      4 -0.28575639
#>  [5,]        6      4  0.05633261
#>  [6,]        3      4 -0.90919539
#>  [7,]        2      4 -0.29754606
#>  [8,]        4      5 -0.30004647
#>  [9,]        1      5 -0.24680336
#> [10,]        7      5  0.92144464
#> [11,]        7      6  0.75174030
#> [12,]        3      6 -0.08901848
#> [13,]        4      6 -0.58308445
#> [14,]        2      6 -0.29491483
(PP <- RotMatPPO(seeds[, 1:7], seeds[, 8], model = "LDA", type = "i-classification"))
#>       Variable Number Coefficient
#>  [1,]        4      1  1.00000000
#>  [2,]        6      2  1.00000000
#>  [3,]        7      3  1.00000000
#>  [4,]        2      4  0.97053457
#>  [5,]        6      4 -0.08108719
#>  [6,]        4      4 -0.01820207
#>  [7,]        3      4  0.22617735
#>  [8,]        7      5  0.72006281
#>  [9,]        5      5  0.23376071
#> [10,]        1      5  0.65334944
#> [11,]        2      6  0.11580946
#> [12,]        7      6 -0.65243552
#> [13,]        1      6 -0.67240009
#> [14,]        5      6  0.32983963

# regression
data(body_fat)
(PP <- RotMatPPO(body_fat[, 2:15], body_fat[, 1], model = "Log", type = "regression"))
#>       Variable Number  Coefficient
#>  [1,]       13      1  1.000000000
#>  [2,]       10      2  1.000000000
#>  [3,]        8      3  1.000000000
#>  [4,]        6      4  1.000000000
#>  [5,]        8      5  0.049250010
#>  [6,]       10      5  0.039119619
#>  [7,]       11      5 -0.098136194
#>  [8,]       14      5  0.238128170
#>  [9,]        7      5 -0.026437207
#> [10,]        1      5 -0.962640528
#> [11,]        9      5 -0.048297436
#> [12,]        6      6 -0.503109602
#> [13,]        2      6 -0.121056500
#> [14,]       12      6 -0.129375503
#> [15,]        5      6  0.572328503
#> [16,]        4      6  0.601174330
#> [17,]        3      6 -0.127645253
#> [18,]       13      6  0.101114936
#> [19,]        1      7 -0.985662862
#> [20,]        4      7  0.012511580
#> [21,]        5      7  0.008120721
#> [22,]        6      7  0.012466368
#> [23,]       14      7  0.152819881
#> [24,]       12      7 -0.068790054
#> [25,]        3      7  0.002199650
(PP <- RotMatPPO(body_fat[, 2:15], body_fat[, 1], model = "Rand", type = "regression"))
#>       Variable Number Coefficient
#>  [1,]        8      1   1.0000000
#>  [2,]        4      2   1.0000000
#>  [3,]       13      3   1.0000000
#>  [4,]        3      4   1.0000000
#>  [5,]        6      5  -0.3779645
#>  [6,]        7      5  -0.3779645
#>  [7,]        5      5   0.3779645
#>  [8,]        1      5   0.3779645
#>  [9,]        2      5   0.3779645
#> [10,]        9      5   0.3779645
#> [11,]       11      5  -0.3779645
#> [12,]       10      6   0.3779645
#> [13,]       13      6   0.3779645
#> [14,]        4      6   0.3779645
#> [15,]        8      6  -0.3779645
#> [16,]       12      6  -0.3779645
#> [17,]       14      6   0.3779645
#> [18,]        3      6  -0.3779645
#> [19,]        6      7  -0.3779645
#> [20,]        7      7  -0.3779645
#> [21,]        5      7   0.3779645
#> [22,]        1      7  -0.3779645
#> [23,]        2      7  -0.3779645
#> [24,]        9      7   0.3779645
#> [25,]       11      7   0.3779645
(PP <- RotMatPPO(body_fat[, 2:15], body_fat[, 1], model = "PPR", type = "regression"))
#>       Variable Number  Coefficient
#>  [1,]        8      1  1.000000000
#>  [2,]       12      2  1.000000000
#>  [3,]        3      3  1.000000000
#>  [4,]        5      4  1.000000000
#>  [5,]       12      5  0.018509419
#>  [6,]        8      5 -0.078601101
#>  [7,]       14      5  0.928046754
#>  [8,]        9      5 -0.077203871
#>  [9,]        6      5 -0.321969779
#> [10,]       11      5  0.146373903
#> [11,]       13      5  0.034032261
#> [12,]        5      6  0.001278698
#> [13,]        2      6  0.015642796
#> [14,]        4      6  0.002157678
#> [15,]       10      6 -0.046318071
#> [16,]        3      6  0.019561921
#> [17,]        1      6 -0.997756843
#> [18,]        7      6 -0.041258474
#> [19,]        1      7 -0.980138044
#> [20,]       14      7  0.166947239
#> [21,]        6      7  0.008437438
#> [22,]       11      7 -0.088693411
#> [23,]        8      7  0.022457497
#> [24,]        9      7 -0.043552373
#> [25,]       10      7  0.033454070