variable importance of oblique decision random forest

Variable importance is computed from permuting OOB data.

Usage

VarImp(forest, X, y)

Arguments

forest

An object of class ODRF.

X

An n by d numerical matrix (preferably) or data frame is used in the ODRF.

y

A response vector of length n is used in the ODRF.

Value

A matrix of importance measure, first column is the predictors and second column is Increased error. Misclassification rate (MR) for classification or mean square error (MSE) for regression. The larger the increased error the more important the variable is.

Details

A note from randomForest package, here are the definitions of the variable importance measures. The measure is computed from permuting OOB data: For each tree, the prediction error on the out-of-bag portion of the data is recorded. Then the same is done after permuting each predictor variable. The difference between the two are then averaged over all trees.

See also

Examples

data(breast_cancer)
set.seed(221212)
train <- sample(1:569, 200)
train_data <- data.frame(breast_cancer[train, -1])
test_data <- data.frame(breast_cancer[-train, -1])

forest <- ODRF(diagnosis ~ ., train_data, split = "gini", parallel = FALSE)
(varimp <- VarImp(forest, train_data[, -1], train_data[, 1]))
#> $varImp
#>                         varible increased_error
#> fractal_dimension_worst      30      0.43312918
#> symmetry_worst               29      0.43154238
#> concave.points_worst         28      0.42876931
#> concavity_worst              27      0.41513632
#> compactness_worst            26      0.38844765
#> smoothness_worst             25      0.38554488
#> area_worst                   24      0.37492995
#> perimeter_worst              23      0.30835416
#> texture_worst                22      0.28817268
#> radius_worst                 21      0.28230358
#> smoothness_se                15      0.22844511
#> area_se                      14      0.22817959
#> compactness_se               16      0.22726991
#> concavity_se                 17      0.22319925
#> symmetry_se                  19      0.22249187
#> concave.points_se            18      0.22241327
#> fractal_dimension_se         20      0.22071566
#> perimeter_se                 13      0.21607246
#> texture_se                   12      0.21075428
#> radius_se                    11      0.20989050
#> symmetry_mean                 9      0.19227163
#> fractal_dimension_mean       10      0.19208747
#> concave.points_mean           8      0.19105162
#> concavity_mean                7      0.17825096
#> compactness_mean              6      0.09275274
#> smoothness_mean               5      0.08253857
#> area_mean                     4      0.07954149
#> perimeter_mean                3      0.06452769
#> texture_mean                  2      0.03472486
#> radius_mean                   1      0.02973542
#> 
#> $split
#> [1] "gini"
#> 
#> attr(,"class")
#> [1] "VarImp"