Samples a p x p uniformly random rotation matrix

Samples a p x p uniformly random rotation matrix via QR decomposition of a matrix with elements sampled iid from a standard normal distribution.

Usage

RandRot(p)

Arguments

p

The columns of an n by p numeric matrix or data frame.

Value

A p x p uniformly random rotation matrix.

See also

Examples

set.seed(220828)
(ODRF:::RandRot(100))
#>                 [,1]          [,2]          [,3]         [,4]          [,5]
#>   [1,]  0.0013126133  0.1581257697 -0.0898401809 -0.123760899  0.0328056565
#>   [2,]  0.2130150326 -0.1174773491 -0.0103198440 -0.059600890  0.2388038759
#>   [3,] -0.0066999139 -0.0645659550 -0.0441180501  0.083019234  0.0461999021
#>   [4,]  0.0214280156  0.0721473618 -0.0996281686 -0.005230795  0.0486488651
#>   [5,]  0.1247601049 -0.0483715834 -0.0146350495  0.213045194 -0.0018697994
#>   [6,] -0.0266064831 -0.0595801217 -0.1024720590 -0.115599156 -0.0847283334
#>   [7,]  0.0873762680 -0.0826533220  0.0076185472 -0.002002670 -0.1841993479
#>   [8,]  0.0307954269  0.1525586181 -0.2232171339 -0.135774295 -0.0380914502
#>   [9,]  0.1055784025 -0.0365169249 -0.0130074347 -0.063709476  0.0022177912
#>  [10,]  0.1454026608 -0.0287819543 -0.0141574689  0.146498899  0.0093128339
#>                 [,6]         [,7]          [,8]          [,9]        [,10]
#>   [1,] -0.0595122049  0.003461576  0.0188957826  1.304059e-01 -0.019857225
#>   [2,]  0.0382448796 -0.076840488 -0.1234678204 -9.360685e-02 -0.080353519
#>   [3,] -0.0057254224 -0.106073954  0.0845534574  7.818394e-02  0.045068973
#>   [4,]  0.1058083790 -0.051300369 -0.0098201483  8.430287e-02  0.035687668
#>   [5,]  0.0373140284 -0.074921415  0.0623254248  7.379306e-02 -0.097176286
#>   [6,] -0.1403199657 -0.075083299 -0.0965007506 -4.852174e-02 -0.015449068
#>   [7,] -0.1260542377 -0.034789706  0.0030078328  3.910404e-02 -0.116259356
#>   [8,] -0.0598031888  0.119406751 -0.0188948694 -3.494230e-02  0.144974285
#>   [9,]  0.0294980776  0.084345783  0.1223325456  1.612050e-02 -0.004310657
#>  [10,] -0.0303685818 -0.025230127  0.0061218455 -5.337848e-02 -0.128633444
#>                [,11]        [,12]         [,13]        [,14]         [,15]
#>   [1,] -0.0264089989  0.001034274  0.0189148607  0.101611306 -1.323729e-01
#>   [2,] -0.1382560688  0.089628460  0.0509280969  0.114465099  3.802055e-02
#>   [3,]  0.1091331691  0.013820037  0.0317725408 -0.018640109 -2.077528e-02
#>   [4,]  0.0524256878 -0.136311993 -0.0287652118  0.173709552 -1.883517e-02
#>   [5,]  0.0401520858  0.030277460 -0.0080846967 -0.139002990 -9.843996e-02
#>   [6,]  0.0615631686  0.091275055 -0.0314736220 -0.094249918 -3.679292e-02
#>   [7,] -0.0224900205 -0.241467393 -0.0601183004 -0.081736006 -1.746361e-01
#>   [8,]  0.1902954639  0.088725182 -0.0632598893 -0.100221401 -8.249249e-02
#>   [9,]  0.0553482901  0.042826327  0.0337909948  0.080004262 -1.096631e-01
#>  [10,]  0.0788333294  0.020294836 -0.0002886304  0.003732852 -1.561243e-01
#>                [,16]        [,17]        [,18]         [,19]         [,20]
#>   [1,] -0.0533688575  0.110473921 -0.071410636  0.0331772400  0.1985598112
#>   [2,]  0.0882699233 -0.039647942 -0.010797261 -0.1621744827 -0.0558684566
#>   [3,] -0.0338422558 -0.102736426  0.106397700  0.0408472269  0.0537643256
#>   [4,] -0.0257243768 -0.033738111 -0.170144304  0.0189384514 -0.0621714743
#>   [5,] -0.0953271079 -0.147200441 -0.053512355  0.0771185018  0.0588851877
#>   [6,]  0.0001510548  0.051885021 -0.104126026  0.1385559173  0.0134271169
#>   [7,]  0.0171203885  0.005265287  0.027864127 -0.1154298830  0.0064065829
#>   [8,]  0.0342524602 -0.031627642 -0.112514772 -0.1449378609  0.1787783469
#>   [9,] -0.0700588889 -0.080769050  0.097738015 -0.1396137223 -0.0530796133
#>  [10,] -0.0872374651  0.142427190 -0.049265916 -0.1125931013 -0.1561866226
#>                [,21]        [,22]         [,23]        [,24]         [,25]
#>   [1,]  0.0839978525 -0.026662934 -0.1255885613 -0.012306052  0.1403564698
#>   [2,]  0.0801415636  0.023066113 -0.0032145942  0.056616680 -0.1448564904
#>   [3,] -0.0705981498 -0.025429768 -0.1030126778  0.023664548 -0.0213384985
#>   [4,]  0.0205387118 -0.094743703  0.0959161770 -0.052786320  0.0495586651
#>   [5,] -0.0072494350  0.027090269 -0.1050206046  0.004657809  0.0003838496
#>   [6,] -0.1120534362 -0.038689840  0.1608433820  0.007910423 -0.0442113945
#>   [7,] -0.0112573792 -0.057956106  0.2082009131  0.164705347  0.1277877553
#>   [8,]  0.0377605639  0.024909187 -0.0347328957  0.040253433 -0.1025776213
#>   [9,]  0.0215219121 -0.041077511  0.0020189538  0.130841287  0.1205189735
#>  [10,] -0.1168328641 -0.067219323 -0.0668432976 -0.020875925 -0.0162452081
#>                [,26]        [,27]        [,28]         [,29]        [,30]
#>   [1,]  0.0145695583  0.110875307  0.024421210  0.2086867259 -0.043591456
#>   [2,]  0.0224940613 -0.088458029  0.070083339  0.1459251278  0.111998972
#>   [3,]  0.0642677672 -0.129740175  0.061907052  0.0111085544  0.043498615
#>   [4,]  0.0533626975 -0.276167445  0.160157292 -0.0141128278  0.007711557
#>   [5,]  0.0885503402  0.130033306 -0.146915097 -0.0293976820  0.038357013
#>   [6,] -0.1361218696  0.056791682  0.108301548  0.0408277845 -0.148811955
#>   [7,] -0.0864183472 -0.046513933  0.027461449 -0.0340313214  0.039109211
#>   [8,] -0.0076565130 -0.129335555 -0.041456727  0.0701128926 -0.045097449
#>   [9,] -0.0734922071  0.017730963 -0.001946285 -0.1272063758  0.089880079
#>  [10,] -0.0368917046 -0.027607590  0.069940749 -0.1012908003 -0.061488572
#>               [,31]         [,32]        [,33]         [,34]        [,35]
#>   [1,] -0.045288375  0.0312275296 -0.057720724 -0.2062562698  0.283882475
#>   [2,]  0.072396957 -0.0237685084  0.178125446 -0.0444867250 -0.081400868
#>   [3,] -0.053065945 -0.1378316334 -0.122578403 -0.0335341833  0.135374230
#>   [4,]  0.059975845  0.0216078523 -0.127850845  0.0515122150 -0.011637962
#>   [5,] -0.024838916  0.1258833210  0.037854831 -0.0873169180 -0.143569867
#>   [6,]  0.244389166  0.0132659629 -0.043790666  0.0263000714  0.098313926
#>   [7,]  0.055889020 -0.2232105434  0.048192264 -0.0077051678  0.027126345
#>   [8,] -0.069780655  0.0380250753  0.150326456 -0.0527295049 -0.119728417
#>   [9,]  0.134045593  0.1878476997 -0.167247482  0.0290985715  0.073319786
#>  [10,] -0.027479421  0.1368693970 -0.129711167 -0.1824609044  0.064003762
#>               [,36]         [,37]        [,38]         [,39]         [,40]
#>   [1,]  0.011162157 -4.420031e-02  0.072042199  0.1860625369  0.1157972926
#>   [2,]  0.081688941  6.420832e-02  0.115595674  0.0551612325  0.1203775933
#>   [3,]  0.055555245 -4.907633e-02 -0.143484864 -0.0920736662  0.1502100138
#>   [4,] -0.187961164  7.324002e-02  0.119697624  0.0185108843  0.0181902026
#>   [5,]  0.049137107 -4.313073e-02  0.080011149  0.1890295893 -0.0451983917
#>   [6,] -0.074300535 -9.899964e-02 -0.027290741 -0.2198682500 -0.0166343828
#>   [7,]  0.005694696  4.003136e-02  0.047030577  0.0266468632  0.0021069684
#>   [8,] -0.064830866  4.353107e-02  0.076077236  0.0694687711  0.0744361594
#>   [9,] -0.123778252 -5.429984e-02 -0.116989201 -0.1100962004  0.0344050068
#>  [10,] -0.007705560  9.440937e-02  0.011130317  0.0360435635 -0.0006452825
#>               [,41]         [,42]         [,43]        [,44]         [,45]
#>   [1,]  0.183248377 -0.0162960855  0.1312753603  0.085482490  1.919686e-02
#>   [2,]  0.049355252 -0.0031798070 -0.0636899482  0.109424356  3.224265e-02
#>   [3,]  0.215909960  0.0380022780 -0.2828208101 -0.106809978 -2.113873e-01
#>   [4,] -0.161554325  0.0253160477  0.0470295953  0.027783232 -7.382287e-03
#>   [5,] -0.246867993  0.0506420959 -0.1093411743  0.187181219  6.571166e-02
#>   [6,] -0.065785426 -0.1136239662 -0.0809869945  0.014344681 -1.219208e-01
#>   [7,] -0.166979037  0.0335104459  0.1233413207 -0.118183175  2.629526e-02
#>   [8,] -0.020501358 -0.0676769491  0.0097974952 -0.109444894 -1.099562e-02
#>   [9,]  0.023964650  0.0661033936 -0.0978025105 -0.029883865  8.055247e-02
#>  [10,]  0.003323702 -0.0499109262  0.0090474346  0.154030344 -2.380586e-02
#>                [,46]         [,47]         [,48]         [,49]         [,50]
#>   [1,]  0.0783184824 -0.1147760947  0.0104617964 -1.014948e-01 -0.0115746974
#>   [2,]  0.1232193456  0.0798170643 -0.0731055674  1.654744e-01  0.0193899404
#>   [3,] -0.0635028443  0.2410569047  0.1010417297  1.530935e-01  0.1510643693
#>   [4,]  0.0309078255  0.0722104595  0.1185279551  1.570290e-02  0.1138758213
#>   [5,] -0.0006204755  0.0817375632  0.1245405511  1.634955e-02 -0.2298028706
#>   [6,] -0.0544515972  0.0724218462  0.0723372498 -5.822824e-02  0.0928655635
#>   [7,]  0.0402930337 -0.0435922012 -0.0504227741  1.974650e-01  0.0560922941
#>   [8,] -0.2332982252  0.0811372569  0.1859794525  9.337578e-02 -0.1693266022
#>   [9,] -0.1044570983 -0.0677520455  0.0137106807  5.046014e-02 -0.1006153151
#>  [10,]  0.0595653346  0.0434370218 -0.0005776567  1.116801e-01  0.0055341928
#>                [,51]        [,52]        [,53]         [,54]         [,55]
#>   [1,] -0.0550736401 -0.072459250  0.076488219 -0.0701951140  0.0578988869
#>   [2,] -0.1794398995  0.004278346  0.003728306 -0.0134039360 -0.0257337933
#>   [3,] -0.0623001987 -0.075990968  0.099036732  0.0710418627  0.0594507413
#>   [4,]  0.1330581466  0.048904577 -0.071306522  0.0595875476 -0.0952943004
#>   [5,] -0.0423594534  0.023776679  0.196189847 -0.0306478232 -0.0688714563
#>   [6,]  0.0642991282  0.009929786  0.055528132  0.0952069031 -0.1812945514
#>   [7,] -0.1413540299 -0.077503061  0.077199383  0.1098194360  0.0570731352
#>   [8,]  0.0959007335 -0.118746643 -0.196788815  0.1082183065 -0.0911055053
#>   [9,]  0.1283769757 -0.091867222 -0.340892498  0.0008364429 -0.0481350146
#>  [10,]  0.0216168491 -0.115651499  0.059732879  0.1387704729 -0.1048971038
#>                [,56]         [,57]        [,58]        [,59]         [,60]
#>   [1,]  0.1294614054 -0.0512171716  0.030622418 -0.137935283  0.1469571215
#>   [2,] -0.1201990243  0.0140517618 -0.229051142 -0.174702273 -0.0812596306
#>   [3,]  0.0239991628 -0.0643372322  0.067923122  0.038284671 -0.2070643458
#>   [4,] -0.1518858809 -0.1454321325  0.144576755  0.059349578 -0.0684321472
#>   [5,]  0.0384149658  0.0547665160 -0.028541830  0.027092737 -0.0033714226
#>   [6,]  0.2574043431 -0.0168525432 -0.138088756 -0.034967883  0.0070964267
#>   [7,]  0.0239461917  0.0425585933 -0.250921001  0.110276673 -0.0725439213
#>   [8,]  0.1533856664  0.1362423783 -0.069211103 -0.022934758 -0.0895989690
#>   [9,] -0.1166047962  0.0334110998  0.073053212 -0.087209748  0.2883415924
#>  [10,]  0.0308937300  0.1281655770  0.010625429  0.044481502  0.1331262989
#>                [,61]         [,62]        [,63]        [,64]        [,65]
#>   [1,] -0.1599620628  0.0962301644  0.057428885 -0.018341240  0.010684465
#>   [2,]  0.0040583232  0.2253860473  0.055644523 -0.112405890  0.084228508
#>   [3,] -0.0406244344 -0.1280392597  0.015334578 -0.030287949 -0.139934757
#>   [4,] -0.1763925531  0.0011560491 -0.088805397 -0.087381320  0.003991745
#>   [5,] -0.2006897238  0.1406187662 -0.059416720  0.125097728 -0.064150640
#>   [6,] -0.0181200150  0.1633801618  0.042845069  0.043254210  0.082050556
#>   [7,] -0.0544474461  0.1442743410 -0.168138301  0.058955416 -0.014661186
#>   [8,]  0.1736660834 -0.1333687426  0.112154244 -0.198761938 -0.004105592
#>   [9,] -0.0443259860  0.1382859651 -0.190020871  0.019055021  0.016500554
#>  [10,]  0.2359195857  0.0920657162  0.087936507  0.098827679  0.151880533
#>                [,66]         [,67]        [,68]         [,69]        [,70]
#>   [1,] -0.0193800745  0.0165826735 -0.078192939  0.0019414933 -0.112200290
#>   [2,] -0.0615558685  0.0296908306 -0.094767840 -0.1304092902  0.027654923
#>   [3,] -0.0552285335  0.0954568668  0.091201871 -0.0009357465 -0.122861418
#>   [4,]  0.0681819958 -0.1686755664  0.103748826  0.1184101169  0.172846508
#>   [5,]  0.1082186591  0.1011955872  0.166201544  0.1698969890 -0.029151199
#>   [6,] -0.1568719828 -0.2362840283 -0.036564254  0.0823118054  0.021100637
#>   [7,] -0.0536322015  0.0907308331  0.089648612 -0.0668343341  0.036684083
#>   [8,]  0.0252137749  0.1284628651 -0.023970531  0.1191147721 -0.065277508
#>   [9,] -0.1365172418  0.1016425456 -0.087895746  0.0600591081 -0.103396180
#>  [10,]  0.0764670568 -0.1407620819 -0.012922711 -0.0959736881 -0.091725224
#>                [,71]         [,72]        [,73]        [,74]        [,75]
#>   [1,]  0.0746545330 -1.661271e-01 -0.100550215  0.183806278 -0.080594475
#>   [2,]  0.1438079632  7.502665e-02  0.093130337 -0.097129207 -0.165679740
#>   [3,]  0.0066628271  3.509118e-02  0.144999197  0.015726417 -0.222548791
#>   [4,]  0.2227828229  2.145115e-02 -0.127418021  0.063631112  0.036768702
#>   [5,] -0.0365344666  2.434657e-01  0.057497800 -0.044666331  0.093734479
#>   [6,]  0.0521272537 -8.826995e-02  0.089076228 -0.134779298  0.073018211
#>   [7,] -0.2672433528 -3.907271e-02 -0.139361485 -0.064055655 -0.025865167
#>   [8,]  0.0002505286  1.201216e-02 -0.057675094 -0.159799295  0.031036037
#>   [9,] -0.1691408100 -1.219036e-02  0.134318005  0.075360944  0.039201976
#>  [10,]  0.0318234790  4.317636e-03  0.069141122 -0.093846893  0.069071262
#>               [,76]         [,77]        [,78]         [,79]         [,80]
#>   [1,]  0.125987307  1.176477e-01  0.190181484  0.0303160076  0.0472069263
#>   [2,]  0.004582028  2.487980e-02  0.060299568 -0.0808395659 -0.0900292359
#>   [3,] -0.061736202  1.215357e-01 -0.035788102 -0.0796993559 -0.0396818843
#>   [4,]  0.209027782  3.640364e-02 -0.012210913 -0.2530876047  0.0827534134
#>   [5,]  0.087883094 -6.183721e-02 -0.098490135 -0.1327825990  0.0443897192
#>   [6,]  0.206117391 -3.249756e-02  0.013361233  0.0607024247 -0.1431944538
#>   [7,] -0.075389728  1.618456e-01  0.050414478  0.0258893938 -0.0541808630
#>   [8,]  0.008860603 -1.761488e-02 -0.074700909 -0.0893126505  0.1129130879
#>   [9,]  0.008085133  1.479409e-01 -0.041443473 -0.0977402715 -0.0818077127
#>  [10,]  0.026570340  1.177848e-02  0.115766577 -0.0796857788  0.0180564435
#>               [,81]         [,82]         [,83]         [,84]         [,85]
#>   [1,] -0.191145431  0.0184236083  1.425186e-01 -0.0946476247 -0.1160954032
#>   [2,]  0.101460984 -0.0136719276 -2.563764e-02  0.0304275754 -0.0559535731
#>   [3,] -0.041761024 -0.0334685751  1.241432e-01 -0.0866840164  0.1316269233
#>   [4,] -0.007652767  0.1262430111 -6.078223e-02  0.1043280118 -0.1866768468
#>   [5,] -0.109723107 -0.0606518498 -7.791521e-03 -0.1145231410 -0.0664941595
#>   [6,]  0.103453717 -0.0367630712 -5.875505e-02 -0.0472213620  0.0849283778
#>   [7,] -0.030303582  0.1016340242  5.096406e-02 -0.0367913350 -0.1605528800
#>   [8,]  0.011180391  0.0102354538  5.312324e-02 -0.0140219812 -0.0316969756
#>   [9,]  0.128729697 -0.0386590151 -3.582298e-02  0.0750976021 -0.0901002177
#>  [10,] -0.085576389  0.1105662994  2.310320e-01  0.1550699359  0.0865264521
#>                [,86]         [,87]        [,88]         [,89]         [,90]
#>   [1,]  0.0846094766  0.0519368109  0.052614628 -0.0227609216  0.1942945775
#>   [2,] -0.0895470443  0.1352505742 -0.046471335 -0.1164271873  0.2101665218
#>   [3,]  0.1141127189 -0.1300493540  0.044079729 -0.0303205271  0.0287750325
#>   [4,]  0.0490292369 -0.0229609682 -0.074909289  0.0482205713  0.0704847613
#>   [5,] -0.0027725678  0.1356134280  0.055742955  0.1107477216  0.0349975445
#>   [6,]  0.0104493670  0.1268962123  0.165172289  0.0654139273 -0.0034247226
#>   [7,]  0.2403312382  0.0501858251 -0.034340537  0.0897955495 -0.0485465650
#>   [8,]  0.1506710517 -0.0291994051 -0.126875771 -0.0612598042  0.0695352916
#>   [9,] -0.0756838720  0.1221534532  0.011236514 -0.2708118002  0.0089905705
#>  [10,]  0.0705459686 -0.2999879317 -0.088485670  0.0038014886 -0.0577511076
#>               [,91]         [,92]        [,93]         [,94]        [,95]
#>   [1,]  0.091278518 -0.0803629478  0.115248002 -0.0640218549  0.022237124
#>   [2,]  0.042370884  0.0047030165 -0.024818137 -0.0062831249  0.112271121
#>   [3,]  0.177855469 -0.0978972102  0.069172118  0.0187590956  0.062510707
#>   [4,]  0.203306517  0.0237843343 -0.130772446 -0.0069092672 -0.022308592
#>   [5,]  0.008055140 -0.0730874694  0.172722112 -0.1163797842  0.059790435
#>   [6,] -0.059948447 -0.1390170397  0.105150133 -0.1586881859  0.107652659
#>   [7,]  0.050187922  0.0023660050  0.003513458  0.1116278367  0.141945671
#>   [8,] -0.023296689  0.0643107946  0.143145753 -0.0688841601  0.050002706
#>   [9,]  0.035841555  0.0328075359  0.152240013  0.0322975746  0.060088043
#>  [10,] -0.005533455  0.0094707387 -0.144483553  0.1157767689  0.021624398
#>               [,96]         [,97]        [,98]         [,99]       [,100]
#>   [1,]  0.049353172  0.0002918049 -0.047715517 -0.0496953104 -0.005833860
#>   [2,] -0.132093428  0.0018635356  0.008437156  0.1971910162 -0.113346908
#>   [3,] -0.060099550  0.1062847484  0.213747048  0.0134758026  0.113224244
#>   [4,] -0.043847116  0.1437351038  0.062785106 -0.0099093183  0.017394397
#>   [5,] -0.001205321 -0.0691835928  0.146301558  0.0851940996  0.077603229
#>   [6,]  0.087118907  0.1461387422  0.079791954  0.1165691216 -0.077231433
#>   [7,]  0.022346423  0.1005403490 -0.078740054 -0.1525223184  0.016769982
#>   [8,] -0.056321954  0.0217421372 -0.231763031 -0.0153204379 -0.027134143
#>   [9,] -0.030070716 -0.0258619337  0.135715705 -0.0341354315  0.099572060
#>  [10,]  0.159168479 -0.1634674385 -0.038986169  0.2069821993  0.219887056
#>  [ reached getOption("max.print") -- omitted 90 rows ]