find best splitting variable and node

A function to select the splitting variables and nodes using one of three criteria.

Usage

best.cut.node(
  X,
  y,
  split,
  lambda = "log",
  weights = 1,
  MinLeaf = 10,
  numLabels = ifelse(split == "mse", 0, length(unique(y)))
)

Arguments

X

An n by d numeric matrix (preferable) or data frame.

y

A response vector of length n.

split

One of three criteria, 'gini': gini impurity index (classification), 'entropy': information gain (classification) or 'mse': mean square error (regression).

lambda

The adjustment parameter of split is used to determine whether to split or not, with the available values being 0, 1 and 'log' (Default).

weights

A vector of values which weigh the samples when considering a split.

MinLeaf

Minimal node size (Default 10).

numLabels

The number of categories.

Value

A list which contains:

BestCutVar: The best split variable.

BestCutVal: The best split points for the best split variable.

BestIndex: Each variable corresponds to the min gini impurity index, the max information gain or the min mean square error.

Examples

### Find the best split variable ###
data(iris)
X <- as.matrix(iris[, 1:4])
y <- iris[[5]]
bestcut <- best.cut.node(X, y, split = "gini")
print(bestcut)
#> $BestCutVar
#> [1] 3
#> 
#> $BestCutVal
#> [1] 2.45
#> 
#> $BestIndex
#> [1] 73.32958 89.61829 54.94401 54.94401
#>