Default values passed to RotMat*

Given the parameter list and the categorical map this function populates the values of the parameter list accoding to our 'best' known general use case parameters.

Usage

defaults(
  paramList,
  split = "entropy",
  dimX = NULL,
  weights = NULL,
  catLabel = NULL
)

Arguments

paramList

A list (possibly empty), to be populated with a set of default values to be passed to a RotMat* function.

split

The criterion used for splitting the variable. 'gini': gini impurity index (classification, default), 'entropy': information gain (classification) or 'mse': mean square error (regression).

dimX

An integer denoting the number of columns in the design matrix X.

weights

A vector of length same as data that are positive weights.(default NULL)

catLabel

A category labels of class list in predictors. (default NULL, for details see Examples of ODT)

Value

Default parameters of the RotMat* function.

dimX An integer denoting the number of columns in the design matrix X.

dimProj Number of variables to be projected, default dimProj="Rand": random from 1 to ncol(X).

numProj the number of projection directions.(default ceiling(sqrt(dimX)))

catLabel A category labels of class list in prediction variables, for details see Examples of ODRF.

weights A vector of length same as data that are positive weights.(default NULL)

lambda Parameter of the Poisson distribution (default 1).

sparsity A real number in \((0,1)\) that specifies the distribution of non-zero elements in the random matrix. When sparsity="pois" means that non-zero elements are generated by the p(lambda) Poisson distribution.

prob A probability \(\in (0,1)\) used for sampling from.

randDist Parameter of the Poisson distribution (default 1).

split The criterion used for splitting the variable. 'gini': gini impurity index (classification, default), 'entropy': information gain (classification) or 'mse': mean square error (regression).

model Model for projection pursuit. (see PPO)

See also

Examples

set.seed(1)
paramList <- list(dimX = 8, numProj = 3, sparsity = 0.25, prob = 0.5)
(paramList <- defaults(paramList, split = "entropy"))
#> $dimX
#> [1] 8
#> 
#> $numProj
#> [1] 3
#> 
#> $sparsity
#> [1] 0.25
#> 
#> $prob
#> [1] 0.5
#> 
#> $dimProj
#> [1] "Rand"
#> 
#> $lambda
#> [1] 1
#> 
#> $randDist
#> [1] "Binary"
#> 
#> $split
#> [1] "entropy"
#> 
#> $model
#> [1] "PPR"
#>